misc adhoc fixes

+++ prisma/migrations/20251012230427_init/migration.sql
@@ -3,8 +3,8 @@ CREATE TABLE "assignment" (
     "id" TEXT NOT NULL,
     "username" TEXT NOT NULL,
     "issued_by" TEXT NOT NULL,
-    "issued_at" BIGINT NOT NULL,
-    "updated_at" BIGINT NOT NULL,
+    "issued_at" INTEGER NOT NULL,
+    "updated_at" INTEGER NOT NULL,
     "issued_status" TEXT NOT NULL,
 
     CONSTRAINT "assignment_pkey" PRIMARY KEY ("id")
